Moved all the routes from index to route files
